Makes GameView.init() accept WordleGameModel
The model is passed into the view, like it shows in
"Head First Design Patterns".

I attempted to change protocol in GameView.swift to specify
init(), but got the following errors:

GameView.swift:9:10: error: keyword 'init' cannot be used as an identifier here
    func init(game: WordleGameModel)
         ^
GameView.swift:9:10: note: if this name is unavoidable, use backticks to escape
it
    func init(game: WordleGameModel)
         ^~~~
         `init`

These errors felt strange and using this did not produce
expected results. I left it as a comment in the protocol.